Razorpay payment link generation is sucessfull

+++ src/main/java/dev/ashish/EcomPaymentService/strategy/PaymentLinkResponse.java
@@ -4,6 +4,6 @@ public interface PaymentLinkResponse {
     String getPaymentLink();
     String getPaymentId();
     String getCurrency();
-    Long getAmount();
+    double getAmount();
     String getDescription();
 }
